+++ test/support/migrations/04_migrate_to_v04.exs
@@ -1,0 +1,7 @@
+defmodule Usher.Test.Repo.Migrations.MigrateToVersion04 do
+  use Ecto.Migration
+
+  def change do
+    Usher.Migration.migrate_to_version("v04")
+  end
+end